Fix 'errors' typo in docs (#22)
The `errors` prop in `useQuery` was returning `undefined` even though I could see the error in the network tab. I couldn't work out why so I tested to see if it should be `error` instead and it worked. I'm not sure if the docs are wrong or if the prop is wrong. Thought I'd send a fix just encase it's the docs.